requirements.txt
argcomplete==1.10.0
asgiref==3.2.5
beautifulsoup4==4.8.0
chardet==3.0.4
Django==3.0.4
django-appconf==1.0.3
django-compressor==2.4
django-extensions==3.0.4
django-sass-processor==0.8
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docx2txt==0.8
EbookLib==0.17.1
entrypoints==0.3
extract-msg==0.23.1
flake8==3.7.9
graphviz==0.14.1
IMAPClient==2.1.0
imutils==0.5.3
libsass==0.19.4
lxml==4.5.1
mccabe==0.6.1
numpy==1.19.0
olefile==0.46
opencv-python==4.2.0.34
pdfminer.six==20181108
Pillow==7.1.2
pycodestyle==2.5.0
pycryptodome==3.9.8
pydotplus==2.0.2
pyflakes==2.1.1
pyparsing==2.4.7
pypiwin32==223
python-docx==0.8.10
python-pptx==0.6.18
pytz==2019.3
pywin32==228
rcssmin==1.0.6
rjsmin==1.1.0
scipy==1.5.1
six==1.12.0
sortedcontainers==2.2.2
soupsieve==2.0.1
sqlparse==0.3.1
tzlocal==1.5.1
xlrd==1.2.0
XlsxWriter==1.2.9
